requirements.txt
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2
ta==0.10.2
jinja2==3.1.2
flask==2.3.3
python-dotenv==1.0.0
line-bot-sdk==3.1.0
requests==2.31.0
schedule==1.2.0
gunicorn==21.2.0 